+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Function `load` will break data consistency.
 - Pressing `ESC` can not skip complex stories with many tags.
 - Function `printItemList` went wrong if the size of item list is bigger than two.
+- Function `make` can not merge items' code in `onMount`/`unMount`
 
 ## [v0.1.1] - 2019-07-18